+++ RPI/Movement.h
@@ -8,7 +8,7 @@ namespace wro
 	{
 		void drive(float speed);// -1 < speed < 1
 		void stop();
-		void steer(BYTE angle);// angle of wheels in degree
+		void steer(short angle);// angle of wheels in degree
 		float getAngle();
 	}
 }
